Integer panel scores and realistic physical covariates in the example
Panel scores are now integers on a 0-10 scale (a panel does not score with
decimals). The covariates use realistic units and ranges with right-skewed
distributions for the skewed quantities, and the spurious proxies are physically
coupled to the mechanistic variables (e.g. refractive index ~ 1.333 + 0.00142 *
Brix, conductivity tracks the acid ions, specific gravity and TDS track Brix),
reported to 3-4 significant figures. To support integer ratings with missing
cells, reshape_to_long and validate_descriptive coerce the score column to a
NaN-aware float. Docs updated to match.

+++ src/process_improve/sensory/ingest.py
@@ -139,7 +139,7 @@ def reshape_to_long(  # noqa: C901, PLR0912, PLR0915
                 **({replicate_col: "replicate"} if replicate_col else {}),
             }
         ).copy()
-        long_df["score"] = pd.to_numeric(long_df["score"], errors="coerce")
+        long_df["score"] = pd.to_numeric(long_df["score"], errors="coerce").astype("float64")
         grand_before = float(np.nanmean(long_df["score"].to_numpy()))
         n_cells_before = int(long_df["score"].notna().sum())
         attr_mean_before = _series_mean_map(long_df, "attribute", "score")
@@ -171,7 +171,7 @@ def reshape_to_long(  # noqa: C901, PLR0912, PLR0915
 
         wide = data.copy()
         for col in value_cols:
-            wide[col] = pd.to_numeric(wide[col], errors="coerce")
+            wide[col] = pd.to_numeric(wide[col], errors="coerce").astype("float64")
         cell_block = wide[value_cols]
         grand_before = float(np.nanmean(cell_block.to_numpy()))
         n_cells_before = int(cell_block.notna().to_numpy().sum())
@@ -203,7 +203,7 @@ def reshape_to_long(  # noqa: C901, PLR0912, PLR0915
         long_df["replicate"] = 1
     for col in ("panelist_id", "product", "attribute"):
         long_df[col] = long_df[col].astype(str).str.strip()
-    long_df["score"] = pd.to_numeric(long_df["score"], errors="coerce")
+    long_df["score"] = pd.to_numeric(long_df["score"], errors="coerce").astype("float64")
 
     # --- Post-reshape marginal aggregates ------------------------------
     grand_after = float(np.nanmean(long_df["score"].to_numpy()))
